# openapi/src/openapi-src.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Your API name",
        "description": "Your API description",
        "version": "0.0.0",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "name": "API Support",
            "email": "your-support-email@11ce303a-bd9d-41c9-8019-2946598c7a07.gov.uk"
        }
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "tags": [{"name": "Foos"}, {"name": "ProgressEntries"}],
    "paths": {
        "/questionnaires/{questionnaireId}/progress-entries": {
            "get": {
                "tags": ["ProgressEntries"],
                "summary": "Get progress entries for a questionnaire.",
                "description": "Gets all progress entries for a questionnaire. The results can be filtered.",
                "operationId": "getProgressEntries",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/questionnaireId"
                    },
                    {
                        "$ref": "#/components/parameters/filter"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "x-scopes": ["read:progress-entries"],
                "x-requests": {
                    "400": {
                        "parameters": [
                            {
                                "name": "filter",
                                "example": {
                                    "NOT-A-VALID-FILTER": "FOO",
                                    "position": "__DELETE__"
                                }
                            }
                        ]
                    },
                    "401": {
                        "auth": false
                    },
                    "403": {
                        "x-scopes": ["create:dummy-resource"]
                    },
                    "404": {
                        "parameters": [
                            {
                                "name": "filter",
                                "example": {
                                    "sectionId": "p--not-a-valid-section",
                                    "position": "__DELETE__"
                                }
                            }
                        ]
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "json-schemas/api/_questionnaires_{questionnaireId}_progress-entries/get/res/200.json"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "type": "progress-entries",
                                            "id": "08dbd0d2-6cee-49e5-a55f-35b31ac4aa9e",
                                            "attributes": {
                                                "sectionId": "p-applicant-declaration"
                                            },
                                            "relationships": {
                                                "section": {
                                                    "data": {
                                                        "type": "sections",
                                                        "id": "p-applicant-declaration"
                                                    }
                                                }
                                            }
                                        }
                                    ],
                                    "links": {
                                        "prev": "questionnaires/{questionnaireId}/progress-entries?page[before]=p-applicant-declaration"
                                    },
                                    "included": [
                                        {
                                            "type": "sections",
                                            "id": "p-applicant-declaration",
                                            "attributes": {
                                                "$schema": "http://json-schema.org/draft-07/schema#",
                                                "type": "object",
                                                "title": "Declaration",
                                                "additionalProperties": false,
                                                "properties": {
                                                    "applicant-declaration": {
                                                        "description": "\n                <p class=\"g-body\">By continuing you confirm that the information you will give is true as far as you know.</p>\n                {{ govukWarningText({\n                    text: \"If you deliberately give false or misleading information, you may get less compensation or be prosecuted.\",\n                    iconFallbackText: \"Warning\"\n                }) }}\n            "
                                                    }
                                                }
                                            },
                                            "relationships": {
                                                "answer": {
                                                    "data": {
                                                        "type": "answers",
                                                        "id": "p-applicant-declaration"
                                                    }
                                                }
                                            }
                                        },
                                        {
                                            "type": "answers",
                                            "id": "p-applicant-declaration",
                                            "attributes": {}
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/foos": {
            "post": {
                "tags": ["Foos"],
                "summary": "Create a new foo resource",
                "description": "Create a foo",
                "operationId": "createFoo",
                "requestBody": {
                    "description": "Some foo data",
                    "required": true,
                    "content": {
                        "application/vnd.api+json": {
                            "schema": {
                                "$ref": "./json-schemas/api/_foos/post/req/201.json"
                            },
                            "example": {
                                "data": {
                                    "type": "foos",
                                    "attributes": {
                                        "bar-attribute": "hello world"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "./json-schemas/api/_foos/post/res/201.json"
                                },
                                "example": {
                                    "data": {
                                        "type": "foos",
                                        "id": "e8ab21fa-4a68-43e8-828c-74eed56c1772",
                                        "attributes": {}
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/foos/{fooId}": {
            "get": {
                "tags": ["Foos"],
                "summary": "Get a foo resource",
                "description": "Gets a foo resource by id",
                "operationId": "getFooById",
                "parameters": [
                    {
                        "name": "fooId",
                        "in": "path",
                        "required": true,
                        "description": "The id of the specific foo instance. Format UUIDv4.",
                        "schema": {
                            "$ref": "./json-schemas/models/definitions/uuid-v4.json"
                        },
                        "example": "285cb104-0c15-4a9c-9840-cb1007f098fb"
                    },
                    {
                        "in": "header",
                        "name": "Accept-Version",
                        "description": "Specify the specific version of the resource (SemVer). Defaults to latest version.",
                        "schema": {
                            "type": "string",
                            "enum": ["1.0.0", "2.0.0"],
                            "default": "2.0.0"
                        },
                        "example": "2.0.0"
                    }
                ],
                "x-scopes": ["read:questionnaires"],
                "x-requests": {
                    "400": {
                        "parameters": [
                            {
                                "name": "questionnaireId",
                                "example": "NOT-A-UUID"
                            }
                        ]
                    },
                    "401": {
                        "auth": false
                    },
                    "404": {
                        "parameters": [
                            {
                                "name": "questionnaireId",
                                "example": "68653be7-877f-4106-b91e-4ba8dac883f4"
                            }
                        ]
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "json-schemas/api/_foos_{fooId}/get/res/200-v1.0.0.json"
                                        },
                                        {
                                            "$ref": "json-schemas/api/_foos_{fooId}/get/res/200-v2.0.0.json"
                                        }
                                    ]
                                },
                                "examples": {
                                    "v1.0.0": {
                                        "summary": "Request Header \"Accept-Version: 1.0.0\"",
                                        "value": {
                                            "data": [
                                                {
                                                    "type": "foos",
                                                    "id": "3cf494a7-9a13-4b67-a4f5-22e6c9dabe56",
                                                    "attributes": {
                                                        "biz-attribute": "No",
                                                        "baz-attribute": true
                                                    }
                                                }
                                            ]
                                        }
                                    },
                                    "v2.0.0": {
                                        "summary": "Request Header \"Accept-Version: 2.0.0\"",
                                        "value": {
                                            "data": [
                                                {
                                                    "type": "foos",
                                                    "id": "3cf494a7-9a13-4b67-a4f5-22e6c9dabe56",
                                                    "attributes": {
                                                        "biz-attribute": false,
                                                        "baz-attribute": true
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "parameters": {
            "questionnaireId": {
                "name": "questionnaireId",
                "in": "path",
                "required": true,
                "description": "The id of the specific questionnaire instance. Format UUIDv4.",
                "schema": {
                    "$ref": "./json-schemas/models/definitions/uuid-v4.json"
                },
                "example": "285cb104-0c15-4a9c-9840-cb1007f098fb"
            },
            "sectionId": {
                "name": "sectionId",
                "in": "path",
                "required": true,
                "description": "The id of the specific questionnaire section.",
                "schema": {
                    "$ref": "./json-schemas/models/definitions/section-id.json"
                },
                "example": "system"
            },
            "filter": {
                "in": "query",
                "name": "filter",
                "description": "Filter progress entries",
                "schema": {
                    "$ref": "./json-schemas/api/parameters/filter.json"
                },
                "style": "deepObject",
                "explode": true,
                "example": {
                    "position": "current"
                }
            },
            "page": {
                "in": "query",
                "name": "page",
                "description": "Paginate progress entries",
                "schema": {
                    "$ref": "./json-schemas/api/parameters/page.json"
                },
                "style": "deepObject",
                "explode": true,
                "example": {
                    "before": "p-applicant-declaration"
                }
            }
        },
        "responses": {
            "BadRequest": {
                "description": "There is an issue with the request",
                "content": {
                    "application/vnd.api+json": {
                        "schema": {
                            "$ref": "./json-schemas/api/responses/400.json"
                        },
                        "example": {
                            "errors": [
                                {
                                    "status": 400,
                                    "title": "400 Bad Request",
                                    "detail": "Request JSON is malformed"
                                }
                            ]
                        }
                    }
                }
            },
            "UnauthorizedError": {
                "description": "Access token is missing or invalid",
                "content": {
                    "application/vnd.api+json": {
                        "schema": {
                            "$ref": "./json-schemas/api/responses/401.json"
                        },
                        "example": {
                            "errors": [
                                {
                                    "status": 401,
                                    "detail": "No authorization token was found"
                                }
                            ]
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "The JWT doesn't permit access to this endpoint",
                "content": {
                    "application/vnd.api+json": {
                        "schema": {
                            "$ref": "./json-schemas/api/responses/403.json"
                        },
                        "example": {
                            "errors": [
                                {
                                    "status": 403,
                                    "title": "403 Forbidden",
                                    "detail": "Insufficient scope"
                                }
                            ]
                        }
                    }
                }
            },
            "NotFound": {
                "description": "The specified resource was not found",
                "content": {
                    "application/vnd.api+json": {
                        "schema": {
                            "$ref": "./json-schemas/api/responses/404.json"
                        },
                        "example": {
                            "errors": [
                                {
                                    "status": 404,
                                    "title": "404 Not Found",
                                    "detail": "Resource /api/v1/questionnaires/2d7caf89-2c72-469f-b19d-17f2a22270b6/sections/answers does not exist"
                                }
                            ]
                        }
                    }
                }
            },
            "Conflict": {
                "description": "Conflict",
                "content": {
                    "application/vnd.api+json": {
                        "schema": {
                            "$ref": "./json-schemas/api/responses/409.json"
                        },
                        "example": {
                            "errors": [
                                {
                                    "status": 409,
                                    "title": "409 conflict",
                                    "detail": "Endpoint /api/v1/questionnaires/68653be7-877f-4106-b91e-4ba8dac883f3/sections/system/answers accepts only resources of type \"answers\""
                                }
                            ]
                        }
                    }
                }
            }
        }
    }
}